Refactoring: Move folder.
